surface_normal/requirements.txt
# Depth Estimation Requirements
# Core dependencies for depth estimation from monocular images

# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0

# Gradio for web interface
gradio>=3.40.0

# Scientific computing
scipy>=1.10.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional: For enhanced visualizations
matplotlib>=3.7.0